Fix ESP32-C3 build. Now project support pre-build binary for additional chip: ESP32-C3, ESP32-S3 and ESP32-S2

+++ main/main.cpp
@@ -3286,7 +3286,7 @@ void loop()
       {
         lastHpSync = millis();
         // If we've retried more than the max number of tries, keep retrying at that fixed interval, which is several minutes.
-        hpConnectionRetries = min(hpConnectionRetries + 1u, HP_MAX_RETRIES);
+        hpConnectionRetries = min((uint32_t)(hpConnectionRetries + 1u), HP_MAX_RETRIES);
         hpConnectionTotalRetries++;
         hp.sync();
       }
